more splitting done
